Device time left (countdown) should be parsed as a double, not an integer.

+++ src/se/markstrom/skynet/skynetremote/xmlparser/ControlXmlParser.java
@@ -38,7 +38,7 @@ public class ControlXmlParser extends XmlParser {
 				device.id = Integer.parseInt(attributes.getNamedItem("id").getFirstChild().getNodeValue());
 				device.name = attributes.getNamedItem("name").getFirstChild().getNodeValue();
 				device.state = Boolean.parseBoolean(attributes.getNamedItem("state").getFirstChild().getNodeValue());
-				device.timeLeft = Integer.parseInt(attributes.getNamedItem("timeleft").getFirstChild().getNodeValue());
+				device.timeLeft = (int) Double.parseDouble(attributes.getNamedItem("timeleft").getFirstChild().getNodeValue());
 				device.type = Integer.parseInt(attributes.getNamedItem("type").getFirstChild().getNodeValue());
 				devices.add(device);
 			}
